requirements.txt
asgiref==3.6.0
Django==4.1.5
django-filter==22.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
drf-writable-nested==0.7.0
pytz==2022.7.1
sqlparse==0.4.3
